services/regulation_service/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pandas
faker
sqlalchemy
psycopg2-binary
httpx